Add strategy for extracting response params from transformers
