+++ config/apidoc.php
@@ -219,6 +219,10 @@ return [
         'bodyParameters' => [
             \Mpociot\ApiDoc\Extracting\Strategies\BodyParameters\GetFromBodyParamTag::class,
         ],
+        'responseParameters' => [
+            \Mpociot\ApiDoc\Extracting\Strategies\ResponseParameters\GetFromResponseParamTag::class,
+            \Mpociot\ApiDoc\Extracting\Strategies\ResponseParameters\GetFromTransformerParamTag::class,
+        ],
         'responses' => [
             \Mpociot\ApiDoc\Extracting\Strategies\Responses\UseTransformerTags::class,
             \Mpociot\ApiDoc\Extracting\Strategies\Responses\UseResponseTag::class,
