+++ src/Extracting/Strategies/Responses/UseTransformerTags.php
@@ -12,6 +12,7 @@ use League\Fractal\Resource\Collection;
 use League\Fractal\Resource\Item;
 use Mpociot\ApiDoc\Extracting\RouteDocBlocker;
 use Mpociot\ApiDoc\Extracting\Strategies\Strategy;
+use Mpociot\ApiDoc\Extracting\TransformerHelpers;
 use Mpociot\ApiDoc\Tools\Flags;
 use Mpociot\ApiDoc\Tools\Utils;
 use Mpociot\Reflection\DocBlock;
@@ -24,6 +25,8 @@ use ReflectionMethod;
  */
 class UseTransformerTags extends Strategy
 {
+    use TransformerHelpers;
+
     /**
      * @param Route $route
      * @param ReflectionClass $controller
@@ -94,108 +97,5 @@ class UseTransformerTags extends Strategy
 
             return null;
         }
-    }
-
-    /**
-     * @param Tag $tag
-     *
-     * @return array
-     */
-    private function getStatusCodeAndTransformerClass($tag): array
-    {
-        $content = $tag->getContent();
-        preg_match('/^(\d{3})?\s?([\s\S]*)$/', $content, $result);
-        $status = $result[1] ?: 200;
-        $transformerClass = $result[2];
-
-        return [$status, $transformerClass];
-    }
-
-    /**
-     * @param array $tags
-     * @param ReflectionMethod $transformerMethod
-     *
-     * @throws Exception
-     *
-     * @return string
-     */
-    private function getClassToBeTransformed(array $tags, ReflectionMethod $transformerMethod): string
-    {
-        $modelTag = Arr::first(array_filter($tags, function ($tag) {
-            return ($tag instanceof Tag) && strtolower($tag->getName()) == 'transformermodel';
-        }));
-
-        $type = null;
-        if ($modelTag) {
-            $type = $modelTag->getContent();
-        } else {
-            $parameter = Arr::first($transformerMethod->getParameters());
-            if ($parameter->hasType() && ! $parameter->getType()->isBuiltin() && class_exists($parameter->getType()->getName())) {
-                // Ladies and gentlemen, we have a type!
-                $type = $parameter->getType()->getName();
-            }
-        }
-
-        if ($type == null) {
-            throw new Exception('Failed to detect a transformer model. Please specify a model using @transformerModel.');
-        }
-
-        return $type;
-    }
-
-    /**
-     * @param string $type
-     *
-     * @return Model|object
-     */
-    protected function instantiateTransformerModel(string $type)
-    {
-        try {
-            // try Eloquent model factory
-
-            // Factories are usually defined without the leading \ in the class name,
-            // but the user might write it that way in a comment. Let's be safe.
-            $type = ltrim($type, '\\');
-
-            return factory($type)->make();
-        } catch (Exception $e) {
-            if (Flags::$shouldBeVerbose) {
-                echo "Eloquent model factory failed to instantiate {$type}; trying to fetch from database.\n";
-            }
-
-            $instance = new $type();
-            if ($instance instanceof IlluminateModel) {
-                try {
-                    // we can't use a factory but can try to get one from the database
-                    $firstInstance = $type::first();
-                    if ($firstInstance) {
-                        return $firstInstance;
-                    }
-                } catch (Exception $e) {
-                    // okay, we'll stick with `new`
-                    if (Flags::$shouldBeVerbose) {
-                        echo "Failed to fetch first {$type} from database; using `new` to instantiate.\n";
-                    }
-                }
-            }
-        }
-
-        return $instance;
-    }
-
-    /**
-     * @param array $tags
-     *
-     * @return Tag|null
-     */
-    private function getTransformerTag(array $tags)
-    {
-        $transformerTags = array_values(
-            array_filter($tags, function ($tag) {
-                return ($tag instanceof Tag) && in_array(strtolower($tag->getName()), ['transformer', 'transformercollection']);
-            })
-        );
-
-        return Arr::first($transformerTags);
     }
 }
